Add external JSON service for About data
